Added example of ADC2 and new function for coulomb repulsion on a grid.
